% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvd_process_subfuncs.R
\name{compute_kl_divergence}
\alias{compute_kl_divergence}
\title{Compute Kullback-Liebler divergence}
\usage{
compute_kl_divergence(frequency_tbl, kl_log_base = "log2")
}
\arguments{
\item{frequency_tbl}{the output of \code{compute_quant_val_dist}}

\item{kl_log_base}{a string indicating the log base to be used in the computations;
defaults to \code{log2}, \code{log} and \code{log10} are also acceptable}
}
\value{
a dataframe with the kullback-liebler divergence value for the site's
distribution compared to the overall, all-site distribution
}
\description{
Compute Kullback-Liebler divergence
}
\keyword{internal}
