% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvd_process_subfuncs.R
\name{compute_quant_val_dist}
\alias{compute_quant_val_dist}
\title{Compute quantitative variable distribution}
\usage{
compute_quant_val_dist(
  qvd_value_file,
  cohort = results_tbl("scd_cohort"),
  grouped_list = "site",
  omop_or_pcornet = "omop",
  time = FALSE
)
}
\arguments{
\item{qvd_value_file}{\emph{tabular input} | dataframe or CSV file with information about each of the variables that should be
examined in the function. contains the following columns:
\itemize{
\item \code{value_name} | \emph{string} |
\item \code{domain_tbl} | \emph{character} | CDM table where the value data is found
\item \code{value_field} | \emph{character} | the name of the field with the quantitative variable OR the name of the person identifier column for patient count checks
\item \code{date_field} | \emph{character} | a date field in the \code{domain_tbl} that should be used for over time analyses
\item \code{concept_field} | \emph{character} | concept_id field with codes from the associated codeset (only needed when codeset is provided)
\item \code{codeset_name} | \emph{character} | the name of the codeset file; DO NOT include the file extension; optional
\item \code{filter_logic} | \emph{character} | a string indicating filter logic that should be applied to achieve the desired variable; optional
}}

\item{cohort}{\emph{tabular input} | A dataframe with the cohort of patients for your study. Should include the columns:
\itemize{
\item \code{person_id} / \code{patid} | \emph{integer} / \emph{character}
\item \code{start_date} | \emph{date}
\item \code{end_date} | \emph{date}
\item \code{site} | \emph{character}
}}

\item{grouped_list}{\emph{vector} | the list of columns that should be used to group the variable table}

\item{omop_or_pcornet}{\emph{string} | Option to run the function using the OMOP or PCORnet CDM as the default CDM
accepts \code{omop} or \code{pcornet}}

\item{time}{\emph{boolean} | a logical that tells the function whether you would like to look at the output over time}
}
\value{
a dataframe with the frequency distribution for each value associated with the variable
of interest and summary statistics (mean, median, q1, q3, sd) describing the distribution
}
\description{
Compute quantitative variable distribution
}
