% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvd_process_subfuncs.R
\name{qvd_euclidean}
\alias{qvd_euclidean}
\title{Euclidean distance for QVD}
\usage{
qvd_euclidean(fot_input_tbl, grp_vars, euclidean_stat = "mean")
}
\arguments{
\item{fot_input_tbl}{the output of \code{compute_fot} when \code{compute_quant_val_dist} is
used as the check_func input}

\item{grp_vars}{list of columns that should be used to group the table}

\item{euclidean_stat}{the statistic that should be computed across all sites to be used
as a comparison to the same site-specific value (i.e. the distance from the site mean to the all site mean)
defaults to \code{mean} but \code{median} is also accepted}
}
\value{
a dataframe with the original stat value, the newly computed all site stat value,
and the euclidean distance between the two distributions stratified by each time point
and the additional grp_vars provided
}
\description{
Euclidean distance for QVD
}
\keyword{internal}
