% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvd_output_subfuncs.R
\name{qvd_ms_anom_cs}
\alias{qvd_ms_anom_cs}
\title{\emph{Multi Site, Anomaly Detection, Cross-Sectional}}
\usage{
qvd_ms_anom_cs(
  process_output,
  value_type_filter = NULL,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{the output of the \code{qvd_process} function}

\item{value_type_filter}{a string or vector of strings to filter the graph to specific variables of interest}

\item{large_n}{a boolean indicating whether the large N visualization, intended for a high
volume of sites, should be used; defaults to FALSE}

\item{large_n_sites}{a vector of site names that can optionally generate a filtered visualization}
}
\value{
a radial lolipop graph displaying the KL divergence value for each site, indicating the divergence from
the all-site frequency distritbution
}
\description{
\emph{Multi Site, Anomaly Detection, Cross-Sectional}
}
