% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvd_output_subfuncs.R
\name{qvd_ms_anom_la}
\alias{qvd_ms_anom_la}
\title{\emph{Multi Site, Anomaly Detection, Longitudinal}}
\usage{
qvd_ms_anom_la(process_output, value_type_filter)
}
\arguments{
\item{process_output}{the output of the \code{qvd_process} function}

\item{value_type_filter}{a string or vector of strings to filter the graph to specific variables of interest}
}
\value{
three graphs:
\enumerate{
\item line graph that shows the smoothed euclidean_stat of a
variable across time computation with the Euclidean distance associated with each line
\item line graph that shows the raw euclidean_stat of a
variable across time computation with the Euclidean distance associated with each line
\item a bar graph with the Euclidean distance value for each site, with the average
euclidean_stat as the fill
}

THIS GRAPH SHOWS ONLY ONE VALUE TYPE AT A TIME!
}
\description{
\emph{Multi Site, Anomaly Detection, Longitudinal}
}
