% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvd_output_subfuncs.R
\name{qvd_ms_exp_cs}
\alias{qvd_ms_exp_cs}
\title{\emph{Multi Site, Exploratory, Cross-Sectional}}
\usage{
qvd_ms_exp_cs(
  process_output,
  value_type_filter = NULL,
  frequency_min = 5,
  display_outliers = FALSE
)
}
\arguments{
\item{process_output}{the output of the \code{qvd_process} function}

\item{value_type_filter}{a string or vector of strings to filter the graph to specific variables of interest}

\item{frequency_min}{an integer to establish a minimum amount of times a value should occur to be included in the output;
aimed at trimming infrequently occurring outliers for a cleaner plot; defaults to 5}

\item{display_outliers}{for boxplot output, a boolean to indicate whether outliers should be displayed; defaults to FALSE}
}
\value{
a plot with boxplots for the distributions at each site, stratified by variable
}
\description{
\emph{Multi Site, Exploratory, Cross-Sectional}
}
