% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvd_output.R
\name{qvd_output}
\alias{qvd_output}
\title{Quantitative Variable Distribution -- Output Generation}
\usage{
qvd_output(
  process_output,
  value_type_filter = NULL,
  frequency_min = 5,
  display_outliers = FALSE,
  summary_stat = "mean"
)
}
\arguments{
\item{process_output}{\emph{tabular output} | the output of the \code{qvd_process} function}

\item{value_type_filter}{\emph{string / vector} | a string or vector of strings to filter the graph to specific variables of interest}

\item{frequency_min}{\emph{integer} | an integer to establish a minimum amount of times a value should occur to be included in the output;
aimed at trimming infrequently occurring outliers for a cleaner plot; defaults to 5}

\item{display_outliers}{\emph{boolean} | for boxplot output, a boolean to indicate whether outliers should be displayed; defaults to FALSE}

\item{summary_stat}{\emph{string} | a string indicating the summary statistic that should be displayed on the plot; required for
exploratory, longitudinal results; defaults to \code{mean}, but \code{median}, \code{q1}, \code{q3}, or \code{sd} are also accepted}
}
\value{
a graph to visualize the results from \code{qvd_process} based on the parameters provided; see documentation
for individual subfunctions for details on specific output
}
\description{
Using the tabular output generated by \code{qvd_process}, this function will build a graph to
visualize the results. Each function configuration will output a bespoke ggplot. Theming can
be adjusted by the user after the graph has been output using \code{+ theme()}. Most graphs can
also be made interactive using \code{make_interactive_squba()}
}
