% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvd_process.R
\name{qvd_process}
\alias{qvd_process}
\title{Quantitative Variable Distribution}
\usage{
qvd_process(
  cohort,
  qvd_value_file,
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  omop_or_pcornet,
  time = FALSE,
  time_span = c("2012-01-01", "2020-01-01"),
  time_period = "year",
  age_groups = NULL,
  sd_threshold = 2,
  kl_log_base = "log2",
  euclidean_stat = "mean"
)
}
\arguments{
\item{cohort}{\emph{tabular input} || \strong{required}

The cohort to be used for data quality testing. This table should contain,
at minimum:
\itemize{
\item \code{site} | \emph{character} | the name(s) of institutions included in your cohort
\item \code{person_id} / \code{patid} | \emph{integer} / \emph{character} | the patient identifier
\item \code{start_date} | \emph{date} | the start of the cohort period
\item \code{end_date} | \emph{date} | the end of the cohort period
}

Note that the start and end dates included in this table will be used to
limit the search window for the analyses in this module.}

\item{qvd_value_file}{\emph{tabular input} || \strong{required}

A dataframe or CSV file with information about each of the variables that should be
examined in the function. Should contain the following columns:
\itemize{
\item \code{value_name} | \emph{string} | a string label for the value variable
\item \code{domain_tbl} | \emph{character} | CDM table where the value data is found
\item \code{value_field} | \emph{character} | the name of the field with the quantitative variable OR the name of the person identifier column for patient count checks
\item \code{date_field} | \emph{character} | a date field in the \code{domain_tbl} that should be used for temporal filtering
\item \code{concept_field} | \emph{character} | the string name of the field in the domain table where the concepts are located (only needed when codeset is provided)
\item \code{codeset_name} | \emph{character} | optional field to include the name of a codeset file
\item \code{vocabulary_field} | \emph{character} | for PCORnet applications, the name of the field in the domain table with a vocabulary identifier to differentiate concepts from one another (ex: dx_type); can be set to NA for OMOP applications
\item \code{filter_logic} | \emph{character} | logic to be applied to the domain_tbl in order to achieve the definition of interest; should be written as if you were applying it in a dplyr::filter command in R
}

To see an example of what this input should look like, see \code{?quantvariabledistribution::qvd_value_file_omop} or
\code{?quantvariabledistribution::qvd_value_file_pcornet}}

\item{multi_or_single_site}{\emph{string} || defaults to \code{single}

A string, either \code{single} or \code{multi}, indicating whether a single-site or
multi-site analysis should be executed}

\item{anomaly_or_exploratory}{\emph{string} || defaults to \code{exploratory}

A string, either \code{anomaly} or \code{exploratory}, indicating what type of results
should be produced.

Exploratory analyses give a high level summary of the data to examine the
fact representation within the cohort. Anomaly detection analyses are
specialized to identify outliers within the cohort.}

\item{omop_or_pcornet}{\emph{string} || \strong{required}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data}

\item{time}{\emph{boolean} || defaults to \code{FALSE}

A boolean to indicate whether to execute a longitudinal analysis}

\item{time_span}{\emph{vector - length 2} || defaults to \code{c('2012-01-01', '2020-01-01')}

A vector indicating the lower and upper bounds of the time series for longitudinal analyses}

\item{time_period}{\emph{string} || defaults to \code{year}

A string indicating the distance between dates within the specified time_span.
Defaults to \code{year}, but other time periods such as \code{month} or \code{week} are
also acceptable}

\item{age_groups}{\emph{tabular input} || defaults to \code{NULL}

If you would like to stratify the results by age group, create a table or
CSV file with the following columns and use it as input to this parameter:
\itemize{
\item \code{min_age} | \emph{integer} | the minimum age for the group (i.e. 10)
\item \code{max_age} | \emph{integer} | the maximum age for the group (i.e. 20)
\item \code{group} | \emph{character} | a string label for the group (i.e. 10-20, Young Adult, etc.)
}

If you would \emph{not} like to stratify by age group, leave as \code{NULL}}

\item{sd_threshold}{\emph{integer} || defaults to \code{2}

An integer indicating the number of standard deviations a value should fall
away from the mean to be considered an outlier. This will be applied to each of
the \verb{Single Site, Anomaly Detection} checks}

\item{kl_log_base}{\emph{string} || defaults to \code{log2}

A string indicating the log base that should be used for the Kullback-Liebler
divergence computation

Acceptable values are: \code{log}, \code{log2}, \code{log10}}

\item{euclidean_stat}{\emph{string} || defaults to \code{mean}

A string indicating the summary statistic that should be used for the
euclidean distance computation in the \verb{Multi-Site, Anomaly Detection, Longitudinal} check

Acceptable values are \code{mean} or \code{median}}
}
\value{
This function will return a dataframe summarizing the
frequency distribution of each quantitative variable. For a
more detailed description of output specific to each check type,
see the PEDSpace metadata repository
}
\description{
This is a plausibility module that will evaluate the distribution of either quantitative
variables (i.e. drug dosages) or the distribution of patient counts (i.e. patients with
inpatient visits). The user will provide definitions for the variables to be
examined (\code{qvd_value_file}). Sample versions of this input are included as
data in the package and are accessible with \verb{quantvariabledistribution::}.
Results can optionally be stratified by site, age group, and/or time. This
function is compatible with both the OMOP and the PCORnet CDMs based on the user's
selection.
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'quantvariabledistribution'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'qvd_process_test',
                      working_directory = my_directory,
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = my_file_folder,
                      cdm_schema = NA)

#' Build mock study cohort
cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = -10000, # RSQLite does not store date objects,
                                      # hence the numerics
                end_date = 30000,
                site = ifelse(person_id \%in\% 1:6, 'synth1', 'synth2'))

#' Create `qvd_value_file` input
qvd_input <- dplyr::tibble('value_name' = c('ibuprofen days supply',
                                            'outpatient visits'),
                           'domain_tbl' = c("drug_exposure",
                                            'visit_occurrence'),
                           'value_field' = c('days_supply',
                                             'person_id'),
                           'date_field' = c('drug_exposure_start_date',
                                            'visit_start_date'),
                           'concept_field' = c('drug_concept_id',
                                               NA),
                           'codeset_name' = c('rx_ibuprofen',
                                              NA),
                           'filter_logic' = c(NA,
                                              'visit_concept_id == 9202'))

#' Execute `qvd_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
qvd_process_example <- qvd_process(cohort = cohort,
                                   multi_or_single_site = 'single',
                                   anomaly_or_exploratory = 'exploratory',
                                   time = FALSE,
                                   omop_or_pcornet = 'omop',
                                   qvd_value_file = qvd_input) \%>\%
  suppressMessages()

qvd_process_example

#' Execute qvd_output` function
qvd_output_example <- qvd_output(process_output = qvd_process_example)

qvd_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(qvd_output_example)
}
