% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvd_output_subfuncs.R
\name{qvd_ss_anom_cs}
\alias{qvd_ss_anom_cs}
\title{\emph{Single Site, Anomaly Detection, Cross-Sectional}}
\usage{
qvd_ss_anom_cs(process_output)
}
\arguments{
\item{process_output}{the output of the \code{qvd_process} function}
}
\value{
a bar plot displaying the proportion of values falling the user-selected number of SD away from the mean
lower and upper outliers are separated by the color of the bar
}
\description{
\emph{Single Site, Anomaly Detection, Cross-Sectional}
}
\keyword{internal}
