% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvd_output_subfuncs.R
\name{qvd_ss_anom_la}
\alias{qvd_ss_anom_la}
\title{\emph{Single Site, Anomaly Detection, Longitudinal}}
\usage{
qvd_ss_anom_la(process_output, value_type_filter = NULL)
}
\arguments{
\item{process_output}{the output of the \code{qvd_process} function}

\item{value_type_filter}{a string or vector of strings to filter the graph to specific variables of interest}
}
\value{
a line plot showing the proportion of values falling the user-selected number of SD away from the mean at
each time point; lower and upper outliers are displayed as separate lines
}
\description{
\emph{Single Site, Anomaly Detection, Longitudinal}
}
