% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvd_output_subfuncs.R
\name{qvd_ss_exp_cs}
\alias{qvd_ss_exp_cs}
\title{\emph{Single Site, Exploratory, Cross-Sectional}}
\usage{
qvd_ss_exp_cs(
  process_output,
  display_outliers = FALSE,
  frequency_min = 5,
  value_type_filter = NULL
)
}
\arguments{
\item{process_output}{the output of the \code{qvd_process} function}

\item{display_outliers}{for boxplot output, a boolean to indicate whether outliers should be displayed; defaults to FALSE}

\item{frequency_min}{an integer to establish a minimum amount of times a value should occur to be included in the output;
aimed at trimming infrequently occurring outliers for a cleaner plot; defaults to 5}

\item{value_type_filter}{a string or vector of strings to filter the graph to specific variables of interest}
}
\value{
a halfeye density plot showing the density of the distribution for each variable with a boxplot underneath
showing the summary of the distribution
}
\description{
\emph{Single Site, Exploratory, Cross-Sectional}
}
