% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvd_output_subfuncs.R
\name{qvd_ss_exp_la}
\alias{qvd_ss_exp_la}
\title{\emph{Single Site, Exploratory, Longitudinal}}
\usage{
qvd_ss_exp_la(process_output, value_type_filter = NULL, summary_stat = "mean")
}
\arguments{
\item{process_output}{the output of the \code{qvd_process} function}

\item{value_type_filter}{a string or vector of strings to filter the graph to specific variables of interest}

\item{summary_stat}{a string indicating the summary statistic that should be displayed on the plot; required for
exploratory, longitudinal results; defaults to \code{mean}, but \code{median}, \code{q1}, \code{q3}, or \code{sd} are also accepted}
}
\value{
a line plot displaying the summary_stat of choice across the time span
}
\description{
\emph{Single Site, Exploratory, Longitudinal}
}
