% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qvd_value_file_pcornet}
\alias{qvd_value_file_pcornet}
\title{QVD Sample Value File (PCORnet)}
\format{
\subsection{qvd_value_file_omop}{

A data frame with 8 columns
\describe{
\item{value_name}{\emph{string} | a string label for the value variable}
\item{domain_tbl}{\emph{character} | CDM table where the value data is found}
\item{value_field}{\emph{character} | the name of the field with the quantitative variable OR the name of the person identifier column for patient count checks}
\item{date_field}{\emph{character} | a date field in the \code{domain_tbl} that should be used for over time analyses}
\item{concept_field}{\emph{character} | concept_id field with codes from the associated codeset (only needed when codeset is provided)}
\item{codeset_name}{\emph{character} | the name of the codeset file; DO NOT include the file extension; optional}
\item{vocabulary_field}{\emph{character} | the name of the column where the vocabulary type associated with the concepts are stored (typically dx_type or px_type, where needed)}
\item{filter_logic}{\emph{character} | a string indicating filter logic that should be applied to achieve the desired variable; optional}
}
}
}
\usage{
qvd_value_file_pcornet
}
\description{
A sample version of the file structure expected for the \code{qvd_value_file}
when the \code{pcornet} CDM is selected. The user should recreate this file and
include their own value definitions.
}
\details{
Files referenced in the \code{codeset_name} column should be kept in the \code{file_subdirectory}
identified when \code{initialize_dq_session} is called.
}
\keyword{datasets}
