% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvd_output.R
\name{qvd_output}
\alias{qvd_output}
\title{Quantitative Variable Distribution -- Output Generation}
\usage{
qvd_output(
  process_output,
  value_type_filter = NULL,
  frequency_min = 5,
  display_outliers = FALSE,
  summary_stat = "mean",
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{\emph{tabular input} || \strong{required}

The tabular output produced by \code{qvd_process}}

\item{value_type_filter}{\emph{string or vector} || defaults to \code{NULL}

A string or vector of strings to filter the graph to specific variables of interest.

A single variable is required to be input in this parameter for the
\verb{Multi-Site, Anomaly Detection, Longitudinal} check. For all others, it is
not required but we recommend limiting to 5 variable or fewer to maintain visibility.}

\item{frequency_min}{\emph{integer} || defaults to \code{5}

An integer indicating the minimum amount of times a value should occur to be
included in the output. This parameter is aimed at trimming infrequently occurring
outliers for a cleaner plot.}

\item{display_outliers}{\emph{boolean} || defaults to \code{FALSE}

For boxplot output, a boolean to indicate whether outliers should be displayed}

\item{summary_stat}{\emph{string} || defaults to \code{mean}

A string indicating the summary statistic that should be displayed on the plot. This is
a required for each \verb{Exploratory, Longitudinal} output.

Acceptable values are \code{mean}, \code{median}, \code{q1}, \code{q3}, or \code{sd}}

\item{large_n}{\emph{boolean} || defaults to \code{FALSE}

For Multi-Site analyses, a boolean indicating whether the large N
visualization, intended for a high volume of sites, should be used. This
visualization will produce high level summaries across all sites, with an
option to add specific site comparators via the \code{large_n_sites} parameter.}

\item{large_n_sites}{\emph{vector} || defaults to \code{NULL}

When \code{large_n = TRUE}, a vector of site names that can add site-level information
to the plot for comparison across the high level summary information.}
}
\value{
This function will produce a graph to visualize the results
from \code{qvd_process} based on the parameters provided. The default
output is typically a static ggplot or gt object, but interactive
elements can be activated by passing the plot through \code{make_interactive_squba}.
For a more detailed description of output specific to each check type,
see the PEDSpace metadata repository
}
\description{
Using the tabular output generated by \code{qvd_process}, this function will build a graph to
visualize the results. Each function configuration will output a bespoke ggplot. Theming can
be adjusted by the user after the graph has been output using \code{+ theme()}. Most graphs can
also be made interactive using \code{make_interactive_squba()}
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'quantvariabledistribution'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'qvd_process_test',
                      working_directory = my_directory,
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = my_file_folder,
                      cdm_schema = NA)

#' Build mock study cohort
cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = -10000, # RSQLite does not store date objects,
                                      # hence the numerics
                end_date = 30000,
                site = ifelse(person_id \%in\% 1:6, 'synth1', 'synth2'))

#' Create `qvd_value_file` input
qvd_input <- dplyr::tibble('value_name' = c('ibuprofen days supply',
                                            'outpatient visits'),
                           'domain_tbl' = c("drug_exposure",
                                            'visit_occurrence'),
                           'value_field' = c('days_supply',
                                             'person_id'),
                           'date_field' = c('drug_exposure_start_date',
                                            'visit_start_date'),
                           'concept_field' = c('drug_concept_id',
                                               NA),
                           'codeset_name' = c('rx_ibuprofen',
                                              NA),
                           'filter_logic' = c(NA,
                                              'visit_concept_id == 9202'))

#' Execute `qvd_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
qvd_process_example <- qvd_process(cohort = cohort,
                                   multi_or_single_site = 'single',
                                   anomaly_or_exploratory = 'exploratory',
                                   time = FALSE,
                                   omop_or_pcornet = 'omop',
                                   qvd_value_file = qvd_input) \%>\%
  suppressMessages()

qvd_process_example

#' Execute qvd_output` function
qvd_output_example <- qvd_output(process_output = qvd_process_example)

qvd_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(qvd_output_example)
}
