% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvd_process.R
\name{qvd_process}
\alias{qvd_process}
\title{Quantitative Variable Distribution}
\usage{
qvd_process(
  cohort,
  qvd_value_file,
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  omop_or_pcornet,
  time = FALSE,
  time_span = c("2015-01-01", "2025-01-01"),
  time_period = "year",
  age_groups = FALSE,
  sd_threshold = 2,
  kl_log_base = "log2",
  euclidean_stat = "mean"
)
}
\arguments{
\item{cohort}{\emph{tabular input} | A dataframe with the cohort of patients for your study. Should include the columns:
\itemize{
\item \code{person_id} / \code{patid} | \emph{integer} / \emph{character}
\item \code{start_date} | \emph{date}
\item \code{end_date} | \emph{date}
\item \code{site} | \emph{character}
}}

\item{qvd_value_file}{\emph{tabular input} | dataframe or CSV file with information about each of the variables that should be
examined in the function. contains the following columns:
\itemize{
\item \code{value_name} | \emph{string} | a string label for the value variable
\item \code{domain_tbl} | \emph{character} | CDM table where the value data is found
\item \code{value_field} | \emph{character} | the name of the field with the quantitative variable OR the name of the person identifier column for patient count checks
\item \code{date_field} | \emph{character} | a date field in the \code{domain_tbl} that should be used for over time analyses
\item \code{concept_field} | \emph{character} | concept_id field with codes from the associated codeset (only needed when codeset is provided)
\item \code{codeset_name} | \emph{character} | the name of the codeset file; DO NOT include the file extension; optional
\item \code{vocabulary_field} | \emph{character} | \emph{PCORNET ONLY} the name of the column where the vocabulary type associated with the concepts are stored (typically dx_type or px_type, where needed)
\item \code{filter_logic} | \emph{character} | a string indicating filter logic that should be applied to achieve the desired variable; optional
}}

\item{multi_or_single_site}{\emph{string} | Option to run the function on a single vs multiple sites
\itemize{
\item \code{single}: run the function for a single site
\item \code{multi}: run the function for multiple sites
}}

\item{anomaly_or_exploratory}{\emph{string} | Option to conduct an exploratory or anomaly detection analysis. Exploratory analyses give a high
level summary of the data to examine the fact representation within the cohort. Anomaly detection
analyses are specialized to identify outliers within the cohort.}

\item{omop_or_pcornet}{\emph{string} | Option to run the function using the OMOP or PCORnet CDM as the default CDM
accepts \code{omop} or \code{pcornet}}

\item{time}{\emph{boolean} | a logical that tells the function whether you would like to look at the output over time}

\item{time_span}{\emph{vector - length 2} | when time = TRUE, this argument defines the start and end dates for the time period of interest. should be
formatted as c(start date, end date) in yyyy-mm-dd date format}

\item{time_period}{\emph{string} | when time = TRUE, this argument defines the distance between dates within the specified time period. defaults
to \code{year}, but other time periods such as \code{month} or \code{week} are also acceptable}

\item{age_groups}{\emph{tabular input} | If you would like to stratify the results by age group,  create a table or CSV file with the following
columns and include it as the \code{age_groups} function parameter:
\itemize{
\item \code{min_age} | \emph{integer} | the minimum age for the group (i.e. 10)
\item \code{max_age} | \emph{integer} | the maximum age for the group (i.e. 20)
\item \code{group} | \emph{character} | a string label for the group (i.e. 10-20, Young Adult, etc.)
}

If you would \emph{not} like to stratify by age group, leave the argument as NULL}

\item{sd_threshold}{\emph{integer} | value indicating the number of standard deviations a value should
fall away from the mean to be considered an outlier; used for single site,
anomaly detection configurations}

\item{kl_log_base}{\emph{string} | string indicating the log base that should be used for the Kullback-Liebler
divergence computation; acceptable values are: \code{log}, \code{log2}, \code{log10}; defaults to \code{log2}}

\item{euclidean_stat}{\emph{string} | string indicating the summary statistic that should be used for the
euclidean distance computation; defaults to \code{mean}, \code{median} is also accepted}
}
\value{
a dataframe with the frequency distribution for each value associated with the variable
of interest and summary statistics (mean, median, q1, q3, sd) describing the distribution
}
\description{
This is a plausibility module that will evaluate the distribution of either quantitative
variables (i.e. drug dosages) or the distribution of patient counts (i.e. patients with
inpatient visits). The user will provide definitions for the variables to be
examined (\code{qvd_value_file}). Sample versions of this input are included as
data in the package and are accessible with \verb{quantvariabledistribution::}.
Results can optionally be stratified by site, age group, and/or time. This
function is compatible with both the OMOP and the PCORnet CDMs based on the user's
selection.
}
